Speed up is_ancestor_of? and branch_contains?
To check if one commit is an ancestor of another - or similarly whether
a commit is on a branch (in effect the same operation, just resolving
the branch name to its HEAD commit), GithubClient was iterating back
through commits from the candidate to see if the ancestor was in the
history.

This was usually fast if it was successful as we're normally deploying
only a few revisions on from the last deployment, so the ancestor (the
last deploy) is very close to the candidate commit. If the commit wasn't
an ancestor, then the commit history would keep being fetched until it
reached the root commit which could take a long time on a large repo.

Now the commit comparison is fetched from GitHub which provides the
merge base commit (see git-merge-base(1)), making the result very easy
to determine as the merge base for two commits where one is the ancestor
of the other will always be the ancestor. If they're not directly
related, e.g. on a different branch, then the merge base will be a third
commit and so the check should fail.

This means the result can be computed with just two API calls and no
iteration over the commit history.

Fixes #11